Remove useless blueprint & api functions
