+++ requirements.txt
@@ -1,6 +1,5 @@
 canonicalwebteam.flask-base==0.3.1
 canonicalwebteam.http==1.0.1
-canonicalwebteam.templatefinder==0.1.2
 Flask==1.1.1
 python-frontmatter==0.4.5
 mistune==0.8.4
